Fix extended FSM method.
Note that the method is prone to failure if the imperative version
of FSM version is used (with exception catching and changing low to
guess. Therefore, it is necessary to instead modify the endpoint of
the feasible bids interval by subtracting a small amount and
re-running the algorithm until it converges.